==> 1.b. Keyword features.ipynb <==
### Preparing PySpark

import os
os.environ['SPARK_HOME'] = '/usr/local/lib/spark'
import findspark
findspark.init()
from pyspark import SparkContext
spark = SparkContext(appName='wordcount')
spark

from pyspark import SQLContext
sqlc = SQLContext(spark)

# reading data
train = sqlc.read.csv("/home/ubuntu/.kaggle/competitions/avito-demand-prediction/train.csv",header=True)

# setting stopwords for russian text
from nltk.corpus import stopwords 
russian_stop = set(stopwords.words('russian'))

# Handle Text with SPARK

train = sqlc.read.csv("/home/ubuntu/.kaggle/competitions/avito-demand-prediction/train.csv",header=True,encoding="utf-8")

# get high deal probability data
trainHDP=train.filter(train["deal_probability"]>0.5).filter(train.description.isNotNull())
# get low deal probability data
trainLDP=train.filter(train["deal_probability"]<0.5).filter(train.description.isNotNull())

# define function to eliminate punctuations
import re
def nopunc(x):
    return re.sub(r'[^\w\s]',' ',x)

### transform text into list, and get list aggragation

mylistH=trainHighDP.rdd \
    .flatMap(lambda x:x["description"].replace(","," ").replace("."," ").split(' ')) \
    .filter(lambda x:x.lower() not in russian_stop) \
    .filter(lambda x: (len(x)>2)) \
    .map(lambda t: (t.lower(), 1)) \
    .reduceByKey(lambda a, b: a + b) \
    .takeOrdered(100, key=lambda pair: -pair[1])

mylistL=trainLowDP.rdd \
    .flatMap(lambda x:x["description"].replace(","," ").replace("."," ").split(' ')) \
    .filter(lambda x:x.lower() not in russian_stop) \
    .filter(lambda x: (len(x)>2)) \
    .map(lambda t: (t.lower(), 1)) \
    .reduceByKey(lambda a, b: a + b) \
    .takeOrdered(100, key=lambda pair: -pair[1])

### get aggragation to a dictionary and get all keywords

mydictH=dict(mylistH)
mydictL=dict(mylistL)

mydictH

dealY=list(mydictH.keys())
dealN=list(mydictL.keys())

#lists of top words 

#top words for deal_probability >= 0.5
dealY = ('состоянии', 'продам', 'хорошем', 'состояние', 'отличном', 'торг', 'очень',     
         'продаю', 'размер', 'цена', 'б/у', 'работает', 'новый', 'хорошее', 'отличное', 
         'телефону', 'года', 'фото', 'новые', 'отдам', 'новая', 'комплекте', 'машина', 
         'продается', 'вопросы', 'цвет', 'ремонт', 'лет', 'квартира', 'идеальном', 
         'квартиру', 'комплект', 'работы', 'авто', 'руб', 'автомобиль', 'рабочем', 
         'коробка', 'телефон', 'срок', 'сдам', 'возможен', 'звоните', 'полностью', 
         'обмен', 'дом', 'коляска', 'велосипед', '"продам', 'двигатель', 'длительный', 
         'срочно', 'подарок', 'имеется', 'самовывоз', 'полный', 'холодильник', 'резина', 
         'музыка', 'время', 'сигнализация', 'диван', 'длина', 'дома', 'салон', 'всё', 
         'связи', 'документы', 'любой', 'отлично', 'рублей', 'высота', 'продаётся', 'чехол', 
         'колеса', 'мальчика', 'большой', 'детский', 'требует', 'рядом', 'стоит', 'девочки', 
         'пробег', 'кожа', 'год', 'фирмы', 'идеальное', 'коляску', 'мало', 'мебель', 'также', 
         'ходу', 'цвета', 'уместен', 'услуги', 'птс', 'торга', 'рост', 'оригинал', 'стол')

#top words for deal_probability < 0.5
dealN = ('состоянии', 'продам', 'размер', 'состояние', 'хорошем', 'очень', 'отличном',   
         'торг', 'продаю', 'новые', 'цена', 'новый', 'отличное', 'новая', 'хорошее', 'платье', 
         'дом', 'б/у', 'квартира', 'цвет', 'продается', 'руб', 'фото', 'рост', 'года', 
         'идеальном', 'квартиру', 'лет', 'рядом', 'кожа', 'телефону', 'длина', 'участок', 
         'цвета', 'мальчика', 'возможен', 'ремонт', 'куртка', 'окна', 'дома', 'вопросы', 
         'девочки', 'работает', 'новое', 'звоните', 'комплект', 'рублей', 'натуральная', 
         'костюм', 'доме', 'носили', 'пару', 'комплекте', 'подарок', 'отдам', 'обмен', 'имеется', 
         'комбинезон', 'туфли', 'телефон', 'девочку', 'идеальное', 'районе', 'пальто', 'кухня', 
         'комнаты', 'уместен', 'наличии', 'внутри', 'документы', 'район', 'стельке', 'продаётся', 
         'большой', 'фирмы', 'школа', 'две', 'срочно', 'полностью', 'мало', 'джинсы', 'вода', 
         'раза', 'качество', 'высота', 'города', 'сапоги', 'ботинки', 'машина', 'гараж', 'покупала', 
         'удобные', 'магазины', 'покупали', 'работы', 'также', 'время', 'газ', 'пишите', 'детский',)

#eliminiate words that are in both lists
dealY_unique = [elem for elem in dealY if elem not in dealN]
dealN_unique = [elem for elem in dealN if elem not in dealY]

#create and apply function to count yes-deal words for all entries in 'description'
wcf = lambda d: [sum(d.count(x) for x in dealY_unique)]
train['wordcountY'] = train['description'].apply(wcf)

#create and apply function to count no-deal words for all entries in 'description'
wcf = lambda d: [sum(d.count(x) for x in dealN_unique)]
train['wordcountN'] = train['description'].apply(wcf)

##### Similarly, by changing file name and path we can do analysis to test data.
