==> 1.d. Text polarity features.ipynb <==
### Preparation

###### This notebook gets the polartity scores from russian text (sentiment analysis).

import pandas as pd
from polyglot.text import Text

mydf=pd.read_csv("testwithimagefeatures.csv",header=0,encoding="utf-8")

# a trial
mytext=Text(mydf["description"][1],hint_language_code="ru")
mytext.polarity

##### define sentiment analysis function

def senti(x):
    try:
        if pd.notnull(x):
            return Text(x,hint_language_code="ru").polarity
        else:
            return -999
    except:
        return -999

##### using multiprocessing to accelerate the process

import numpy as np
from multiprocessing import Pool
import time

time1=time.time()

# Number of CPU cores
cores = 32
# Define partitions for data
partitions = cores

def parallelize_dataframe(data, func):
        print ("parallelizing dataframe")
        data_split = np.array_split(data, partitions)
        pool = Pool(cores)
        data = pd.concat(pool.map(func, data_split))
        pool.close()
        pool.join()
        return data

def test_func(data):
    print ("applying function")
    data["descsentiment"] = data["description"].apply(senti)
    return data
    
mydf = parallelize_dataframe(mydf, test_func)

print ("running for ",round((time.time()-time1)/60)," mins")

##### do analysis to both title and description

def test_func(data):
    print ("applying function")
    data["titlesentiment"] = data["title"].apply(senti)
    return data
    
mydf = parallelize_dataframe(mydf, test_func)

print ("running for ",round((time.time()-time1)/60)," mins")

##### checking results

mydf.loc[:,["titlesentiment","descsentiment"]].head(50)

##### output csv

mydf.to_csv("testwithsentiment.csv",header=True,index=False,encoding="utf-8")
print ("completed")

# silimarly we can change the path and file name to analyze both training and test data
