==> 1.e. Geographic features.ipynb <==
##### the additional regional and city data is available in Kaggle Discussion Board.
##### All these data are public available and are allowed to be used in this competition

# Regional Information

import pandas as pd
import numpy as np

df_train=pd.read_csv("testwithsentiment.csv",header=0,encoding="utf-8")

##### create a russion translation for the data file is English version

region_map = {"Свердловская область" : "Sverdlovsk oblast",
            "Самарская область" : "Samara oblast",
            "Ростовская область" : "Rostov oblast",
            "Татарстан" : "Tatarstan",
            "Волгоградская область" : "Volgograd oblast",
            "Нижегородская область" : "Nizhny Novgorod oblast",
            "Пермский край" : "Perm Krai",
            "Оренбургская область" : "Orenburg oblast",
            "Ханты-Мансийский АО" : "Khanty-Mansi Autonomous Okrug",
            "Тюменская область" : "Tyumen oblast",
            "Башкортостан" : "Bashkortostan",
            "Краснодарский край" : "Krasnodar Krai",
            "Новосибирская область" : "Novosibirsk oblast",
            "Омская область" : "Omsk oblast",
            "Белгородская область" : "Belgorod oblast",
            "Челябинская область" : "Chelyabinsk oblast",
            "Воронежская область" : "Voronezh oblast",
            "Кемеровская область" : "Kemerovo oblast",
            "Саратовская область" : "Saratov oblast",
            "Владимирская область" : "Vladimir oblast",
            "Калининградская область" : "Kaliningrad oblast",
            "Красноярский край" : "Krasnoyarsk Krai",
            "Ярославская область" : "Yaroslavl oblast",
            "Удмуртия" : "Udmurtia",
            "Алтайский край" : "Altai Krai",
            "Иркутская область" : "Irkutsk oblast",
            "Ставропольский край" : "Stavropol Krai",
            "Тульская область" : "Tula oblast"}

df_train['region_en'] = df_train['region'].apply(lambda x : region_map[x])

# get data from additional csv
regional = pd.read_csv("regional.csv", index_col = [0])

rDense = regional["Density_of_region(km2)"]
rRural = regional["Rural_%"]
rTime_zone = regional["Time_zone"]
rPopulation = regional["Total_population"]
rUrban = regional["Urban%"]

# preparing dataframe
reg_index = np.array([regional.index[i].lower() for i in range(len(regional))])
rDense.index = reg_index
rRural.index = reg_index
rTime_zone.index = reg_index
rPopulation.index = reg_index
rUrban.index = reg_index

#### assign regional information to dataset

df_region = df_train["region_en"]

reg_dense = np.array([rDense[df_region[i].lower()] for i in range(len(df_train))])
reg_rural = np.array([rRural[df_region[i].lower()] for i in range(len(df_train))])
reg_Time_zone = np.array([rTime_zone[df_region[i].lower()] for i in range(len(df_train))])
reg_Population = np.array([rPopulation[df_region[i].lower()] for i in range(len(df_train))])
reg_Urban = np.array([rUrban[df_region[i].lower()] for i in range(len(df_train))])

df_train["reg_dense"] = reg_dense
df_train["rural"] = reg_rural
df_train["reg_Time_zone"] = reg_Time_zone
df_train["reg_Population"] = reg_Population
df_train["reg_Urban"] = reg_Urban

##### check results

df_train.head(5)

##### output file

df_train.to_csv("testwithsentiment.csv",header=True,index=False,encoding="utf-8")
print ("completed")

# City Population

##### Similarly, use additional information to create city population feature

df_train=pd.read_csv("testwithsentiment.csv",header=0,encoding="utf-8")

#### the file is in a different format, so read it as dictionary

import csv
with open('city_population.csv',encoding="utf-8") as f:
    reader = csv.DictReader(f)
    city_list={}
    for row in reader:
        mykey=row["city"]
        myvalue=row["population"]
        city_list[mykey] = myvalue

##### define function to assign values. Note that some cities are missing.

def citypopu(x):
    try:
        return float(city_list[x])
    except:
        return -999

df_train['city_population'] = df_train['city'].apply(citypopu)

##### check missing values and results

len(df_train[df_train.city_population<0])

df_train.head(5)

##### output file

df_train.to_csv("testwithsentiment.csv",header=True,index=False,encoding="utf-8")
print ("completed")
