==> 4.a. Blended model.ipynb <==
#### Preparation

import pandas as pd
import numpy as np

### get all models we have

testdata1=pd.read_csv('lgsubcat.csv',header=0)
testdata2=pd.read_csv('rnn224.csv',header=0)
testdata3=pd.read_csv('xgb.csv',header=0)
testdata4=pd.read_csv('lgsub2219.csv',header=0)
testdata5=pd.read_csv('StackTest.csv',header=0)
testdata5["deal_probability"]=testdata5["lgb_pred"]
testdata6=pd.read_csv('ridge.csv',header=0)

### blend with a weighted average

# round all numbers to four digits
testdata1["deal_probability"]=round(testdata1["deal_probability"],4)
testdata2["deal_probability"]=round(testdata2["deal_probability"],4)
testdata3["deal_probability"]=round(testdata3["deal_probability"],4)
testdata4["deal_probability"]=round(testdata4["deal_probability"],4)
testdata5["deal_probability"]=round(testdata5["deal_probability"],4)
testdata6["deal_probability"]=round(testdata6["deal_probability"],4)

#### the weight is based on experience and model rmse (lower rmse, higher weight)

preds=0.05*testdata1["deal_probability"]+0.25*testdata2["deal_probability"]\
    +0.1*testdata3["deal_probability"]+0.35*testdata4["deal_probability"]\
    +0.15*testdata5["deal_probability"]+0.1*testdata6["deal_probability"]
    
preds=round(preds,4)

output = pd.DataFrame({'deal_probability': preds, 'item_id': testdata1['item_id']})
output = output[['item_id', 'deal_probability']]

### considering RMSE scoring, try to narrow the data range to lower RMSE

def mycenter(x):
    if x >= 0.9999:
        x=0.9999
    elif x <= 0.0001:
        x=0.0001
    else : 
        x=x
    return x

output["deal_probability"]=output["deal_probability"].apply(mycenter)

### output and submit

output.to_csv('myblending.csv', index=False) 

### Final Score:
#### Public Score: 0.2205
#### Private Score: 0.2243
